feat: skill verification request notification
