+++ src/hooks/useThemeColor.ts
@@ -1,0 +1,12 @@
+/** @format */
+
+import { FC } from "react";
+import { useTheme } from "next-themes";
+
+const UseThemeColor = () => {
+  const { theme, setTheme } = useTheme();
+
+  return theme;
+};
+
+export default UseThemeColor;
