+++ src/mobx/stores/index.ts
@@ -1,7 +1,5 @@
 /** @format */
 
-"use client";
-
 import { Instance, types } from "mobx-state-tree";
 import { persist } from "mst-persist";
 import localForage from "localforage";
@@ -13,6 +11,7 @@ import { FanStore } from "./fans";
 import { AthleteStore } from "./athletes";
 import { SchoolStore } from "./schools";
 import { SkillTypeStore } from "./skill-types";
+import { VerificationRequestStore } from "./verification-request";
 
 export const RootStore = types.model("RootStore", {
   counterStore: CounterStore,
@@ -23,6 +22,7 @@ export const RootStore = types.model("RootStore", {
   fanStore: FanStore,
   schoolStore: SchoolStore,
   SkillTypeStore :SkillTypeStore,
+  verificationRequestStore:VerificationRequestStore
 });
 
 export interface RootStore extends Instance<typeof RootStore> {}
@@ -38,6 +38,7 @@ export const createRootStore = () => {
       fanStore: FanStore.create(),
       schoolStore: SchoolStore.create(),
       SkillTypeStore :SkillTypeStore.create(),
+      verificationRequestStore:VerificationRequestStore.create()
 
     },
     { NODE_ENV: process.env.NODE_ENV }
