+++ src/services/graphql/queries/coach.graphql
@@ -122,6 +122,7 @@ query getCoach($where: CoachProfileWhereUniqueInput!) {
       avatar
       isActive
       email
+      dob
       accountType {
         id
         title
